Tested barge_in failure on left
Updated test/sim so that robot tries to barge in to the left.
Still exhibits qualitatively reasonable behavior. However, the
rewards seemed to be less stable and lower overall. Additionally,
the robot did not move as far away from the "human" obstacles as
when the obstacles were on the right.