Dynamic training and testing
Trained a model on a dynamic_barge_in_right scenario. Performs reasonably
well overall. Also tested a model trained on a static scene as well as
one trained on a random scene against the dynamic barge in scenarios. Both
produced reasonable behavior.

Also tested the model trained on a dynamic scene against a static scene
on which it performed well.

Note, also reorganized the tests folder to split up barge, dynamic_barge,
and random tests. The name of the folder corresponds to the scene on
which the model being tested was trained. So everything in the `random`
folder are results produced by a model trained on a random scene, but
is not necessarily itself a random scene.