Debugging loss function
Tried modifying the loss function so that it didn't explode in the
negative direction. Got a model to converge but upon testing it
simply predicts huge variances with mus that are relatively
far from the true values (but the huge variances mean the
probability of observing the training data is still relatively high).
Working on this.