Debugging and documentation
Debugged addition of forward simulation for training a model on
a given scenario. I.e. a path to a success controller can be passed
and this model's predictions will be followed until it is determined
that its predictions are faulty. Determining when the predictions
are incorrect has proven more difficult.

Added documentation to the functions necessary to make this change.