Updated prediction training data
Moved to batches of 1 so that training data for prediction model
can include all occupancy maps, not just the one for the robot.
This is what the original network was designed to do, but it
could batch because the number of humans in a given episode
was constant. This is not true across episodes, which is why the
predictor must have batches of 1.