Updated generation of training data
The training data generator now also stores the initial position and
radii of all the agents in the simulation as well as the list of
obstacles for visualization purposes later on.

The simulator also was modified so that `update_visualization` returns
the string it writes to the output file if there is one.