Tested with rolling average window of size 1
This allows the robot to respond more quickly to unfamiliar scenes and it
is effective. Surprisingly, in 8/10 test cases on a familiar scene, the robot
continued with the success controller the entire time (without switching over
to the failure controller). In the two cases in which it did switch, the switch
occurred past the halfway mark (time steps 77 and 59 respectively). The failure
controller, which was poorly trained, also achieved a relatively good reward
indicating that it can perform reasonable actions even if it is not in a failure
situation.