Best for failure prediction
These two models (see test.sh) did the best for separating when the robot
was failing from when it was succeeding. For some reason, having trouble reproducing.
Perhaps it was from having the same train and validation sets? I don't see why that
would be the case but that is the main thing I changed. The validation set should
have no impact on the training though, it should just be an indicator.